+++ test/supports/app_run_tests/app_with_lookups.py
@@ -1,7 +1,7 @@
 # Copyright Modal Labs 2022
 import modal
 
-app = modal.App("my-app")
+app = modal.App("my-app", include_source=True)  # TODO: remove include_source=True)
 
 nfs = modal.NetworkFileSystem.from_name("volume_app").hydrate()