+++ test/webhook_test.py
@@ -13,7 +13,7 @@ from modal.functions import Function
 from modal.running_app import RunningApp
 from modal_proto import api_pb2
 
-app = App()
+app = App(include_source=True)  # TODO: remove include_source=True when automount is disabled by default
 
 
 @app.function(cpu=42)